adding coercer for default values/secrets #108 - split validator creation into separate module
